+++ xss-clean-spring/src/main/java/io/github/haiphamcoder/xss/config/XssProperties.java
@@ -5,7 +5,7 @@ import java.util.Map;
 
 import org.springframework.boot.context.properties.ConfigurationProperties;
 
-@ConfigurationProperties(prefix = "xss")
+@ConfigurationProperties(prefix = "xss.cleaner")
 public class XssProperties {
 
     /**